Updating variable types and adding the first model

+++ Assets/Code/Domain/Opinion.cs
@@ -6,15 +6,15 @@
 public class Opinion
 {
     /// <summary>
-    /// Gets the X coordinate.
+    /// Gets the X coordinate. It represents a value from -1 to 1 for each choice.
     /// </summary>
-    public List<int> OpinionVector { get; }
+    public List<float> OpinionVector { get; }
 
     /// <summary>
     /// Initializes a new instance of the <see cref="Opinion"/> class.
     /// </summary>
     /// <param name="opinionVector">Represents the opinion vector.</param>
-    public Opinion(List<int> opinionVector)
+    public Opinion(List<float> opinionVector)
     {
         OpinionVector = opinionVector;
     }
